# server/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Record_string.string_": {
				"properties": {},
				"additionalProperties": {
					"type": "string"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"UserRole": {
				"type": "string",
				"enum": [
					"ADMIN",
					"PROJECT_MANAGER",
					"COLLABORATOR"
				]
			},
...
			"ApiResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"_36_Enums.Priority": {
				"type": "string",
				"enum": [
					"LOW",
					"MEDIUM",
					"HIGH"
				]
			},
			"_36_Enums.Status": {
				"type": "string",
				"enum": [
					"TODO",
					"IN_PROGRESS",
					"COMPLETED"
				]
			},
			"DefaultSelection_Prisma._36_TaskPayload_": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"position": {
						"type": "number",
						"format": "double"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.Status"
					},
					"priority": {
						"$ref": "#/components/schemas/_36_Enums.Priority"
					},
					"dueDate": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"projectId",
					"updatedAt",
					"createdAt",
					"position",
					"tags",
					"status",
					"priority",
					"dueDate",
					"description",
					"title",
					"id"
				],
				"type": "object"
			},
			"Task": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_TaskPayload_",
				"description": "Model Task"
			},
			"ApiResponse_Task-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array",
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommentAuthor": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Attachment": {
				"properties": {
					"id": {
						"type": "string"
					},
					"filename": {
						"type": "string"
					},
					"fileKey": {
						"type": "string"
					},
					"presignedUrl": {
						"type": "string"
					},
					"mimeType": {
						"type": "string"
					},
					"fileSize": {
						"type": "number",
						"format": "double"
					},
					"taskId": {
						"type": "string"
					},
					"commentId": {
						"type": "string",
						"nullable": true
					},
					"uploadedById": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"id",
					"filename",
					"fileKey",
					"mimeType",
					"fileSize",
					"taskId",
					"commentId",
					"uploadedById",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Comment": {
				"properties": {
					"id": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"taskId": {
						"type": "string"
					},
					"authorId": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/CommentAuthor"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/Attachment"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"content",
					"taskId",
					"authorId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TaskAssignee": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"assignedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"userId",
					"name",
					"email",
					"assignedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SharedTask": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"dueDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"priority": {
						"type": "string",
						"enum": [
							"LOW",
							"MEDIUM",
							"HIGH"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"TODO",
							"IN_PROGRESS",
							"COMPLETED"
						]
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"position": {
						"type": "number",
						"format": "double"
					},
					"projectId": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"comments": {
						"items": {
							"$ref": "#/components/schemas/Comment"
						},
						"type": "array"
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/Attachment"
						},
						"type": "array"
					},
					"assignees": {
						"items": {
							"$ref": "#/components/schemas/TaskAssignee"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title",
					"priority",
					"status",
					"tags",
					"position",
					"projectId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_SharedTask_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/SharedTask"
							}
						],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Task_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Task"
							}
						],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTaskRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"projectId": {
						"type": "string"
					},
					"dueDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"priority": {
						"type": "string",
						"enum": [
							"LOW",
							"MEDIUM",
							"HIGH"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"TODO",
							"IN_PROGRESS",
							"COMPLETED"
						]
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"position": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"projectId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTaskRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"dueDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"priority": {
						"type": "string",
						"enum": [
							"LOW",
							"MEDIUM",
							"HIGH"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"TODO",
							"IN_PROGRESS",
							"COMPLETED"
						]
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"position": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ProjectStatus": {
				"type": "string",
				"enum": [
					"ACTIVE",
					"ARCHIVED",
					"COMPLETED"
				]
			},
			"Project": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"$ref": "#/components/schemas/ProjectStatus"
					},
					"endDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						],
						"nullable": true
					},
					"ownerId": {
						"type": "string"
					},
					"createdAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"updatedAt": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"status",
					"endDate",
					"ownerId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Project_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Project"
							}
						],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_Comment-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Comment"
						},
						"type": "array",
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Comment_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Comment"
							}
						],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateAttachmentRequest": {
				"properties": {
					"filename": {
						"type": "string"
					},
					"fileKey": {
						"type": "string"
					},
					"mimeType": {
						"type": "string"
					},
					"fileSize": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"filename",
					"fileKey",
					"mimeType",
					"fileSize"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCommentRequest": {
				"properties": {
					"content": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/CreateAttachmentRequest"
						},
						"type": "array"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthData": {
				"properties": {
					"token": {
						"type": "string"
					},
					"mustResetPassword": {
						"type": "boolean"
					}
				},
				"required": [
					"token",
					"mustResetPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_AuthData_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AuthData"
							}
						],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordRequest": {
				"properties": {
					"currentPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					},
					"confirmNewPassword": {
						"type": "string"
					}
				},
				"required": [
					"currentPassword",
					"newPassword",
					"confirmNewPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Attachment-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Attachment"
						},
						"type": "array",
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Attachment_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Attachment"
							}
						],
						"nullable": true
					},
					"errors": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.string_"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"errors"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "@nextask/server",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/tasks": {
			"get": {
				"operationId": "GetTasks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "projectId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"TODO",
								"IN_PROGRESS",
								"COMPLETED"
							]
						}
					},
					{
						"in": "query",
						"name": "priority",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"LOW",
								"MEDIUM",
								"HIGH"
							]
						}
					},
					{
						"in": "query",
						"name": "tags",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					}
				]
			},
			"post": {
				"operationId": "CreateNewTask",
				"responses": {
					"201": {
						"description": "Task Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": [
							"project:manager"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskRequest"
							}
						}
					}
				}
			}
		},
		"/tasks/{id}": {
			"get": {
				"operationId": "GetTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_SharedTask_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateExistingTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Task_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTaskRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteExistingTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [
					{
						"jwt": [
							"project:manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/projects/{id}/complete": {
			"patch": {
				"operationId": "Complete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Project_"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": [
							"project:manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/tasks/{taskId}/comments": {
			"get": {
				"operationId": "GetComments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Comment-Array_"
								}
							}
						}
					}
				},
				"description": "Retrieves all comments for a specific task.",
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Comment_"
								}
							}
						}
					}
				},
				"description": "Posts a new comment to a specific task.",
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentRequest"
							}
						}
					}
				}
			}
		},
		"/comments/{commentId}": {
			"delete": {
				"operationId": "RemoveComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Deletes a comment by ID.",
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "commentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AuthData_"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequest"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AuthData_"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest"
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_AuthData_"
								}
							}
						}
					}
				},
				"description": "First-login password reset (and any subsequent voluntary reset).\r\nRequires a valid JWT — users with mustResetPassword=true are ONLY\r\nallowed to call this endpoint until they complete the reset.",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordRequest"
							}
						}
					}
				}
			}
		},
		"/tasks/{taskId}/attachments": {
			"get": {
				"operationId": "GetAttachments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Attachment-Array_"
								}
							}
						}
					}
				},
				"description": "Retrieves all attachments for a specific task.",
				"tags": [
					"Attachments"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "UploadAttachmentMetadata",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Attachment_"
								}
							}
						}
					}
				},
				"description": "Uploads file metadata directly to a task.",
				"tags": [
					"Attachments"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAttachmentRequest"
							}
						}
					}
				}
			}
		},
		"/attachments/{attachmentId}": {
			"delete": {
				"operationId": "RemoveAttachment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Deletes an attachment metadata record by ID.",
				"tags": [
					"Attachments"
				],
				"security": [
					{
						"jwt": [
							"project:member"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "attachmentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}